requirements.txt
chevron
CherryPy==18.5.0
cryptography
dnspython
docker
GeoIP
graypy
iso8601
jira==2.0.0
kubernetes
netaddr
PyJWT==2.0.0
pymongo==3.6.1
python-dateutil==2.8.1
python-geoip-python3==1.3
python-ldap==3.2.0
python-memcached
PyYAML
requests
rsa==4.0
shellescape
